Move go code to pkg/libvirt/
This will be more consistant with what other go packages are doing,
having go files in the top-level directory is very unusual
